requirements.txt
django-test-plus
aiodns==3.0.0
aiosmtpd==1.4.4.post2
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
atpublic==4.0
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.3
django-allauth==0.54.0
django-allauth-ui==0.1.4
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-invitations==2.0.0
django-rest-framework==0.1.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
flower==2.0.0
humanize==4.7.0
idna==3.4
inflection==0.5.1
kombu==5.3.1
oauthlib==3.2.2
packaging==23.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycares==4.3.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tornado==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
uuid==1.30
verify-email==2.4.3
vine==5.0.0
wcwidth==0.2.6
gunicorn==21.2.0